05_tool_calling/requirements.txt
langchain
langchain_openai
langgraph
python-dotenv
langchain_community
tavily-python
requests


